fix the jank
